Revert "bg_atlasapi -> brainglobe_atlasapi, package restructure"
This reverts commit 120c3b2c183060d698105a2820d33447a5eced04.
