+++ .gitignore
@@ -1,4 +1,4 @@
-brainglobe_atlasapi/bg_config.conf
+bg_atlasapi/bg_config.conf
 
 workspace.py
 .vscode/
