+++ bg_atlasapi/descriptors.py
