+++ bg_atlasapi/list_atlases.py
@@ -6,7 +6,7 @@ from rich import print as rprint
 from rich.panel import Panel
 from rich.table import Table
 
-from brainglobe_atlasapi import config, descriptors, utils
+from bg_atlasapi import config, descriptors, utils
 
 
 def get_downloaded_atlases():
