+++ bg_atlasapi/structure_tree_util.py
