+++ tests/atlasapi/conftest.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import pytest
 
-from brainglobe_atlasapi.bg_atlas import BrainGlobeAtlas
+from bg_atlasapi.bg_atlas import BrainGlobeAtlas
 
 
 @pytest.fixture()
