+++ tests/atlasapi/test_list_atlases.py
@@ -1,4 +1,4 @@
-from brainglobe_atlasapi.list_atlases import (
+from bg_atlasapi.list_atlases import (
     get_atlases_lastversions,
     get_downloaded_atlases,
     get_local_atlas_version,
