+++ tests/atlasapi/test_structure_dict.py
@@ -1,9 +1,9 @@
 import meshio as mio
 import pytest
 
-from brainglobe_atlasapi import descriptors
-from brainglobe_atlasapi.structure_class import StructuresDict
-from brainglobe_atlasapi.utils import read_json
+from bg_atlasapi import descriptors
+from bg_atlasapi.structure_class import StructuresDict
+from bg_atlasapi.utils import read_json
 
 structures_list = [
     {
